hoàn thành đô án

+++ Modules/Admin/Http/Controllers/AdminCategoryController.php
@@ -7,9 +7,16 @@ use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
 use App\Http\Requests\RequestCategory;
 use App\Models\Category;
-class AdminCategoryController extends Controller
+use App\Models\Product;
+
+class AdminCategoryController extends AdminHeaderController
 {
-   
+
+    public function __construct()
+	{
+		parent::__construct();
+	}
+
     public function index()
     {
         $categories = Category::select('id','c_name','c_title_seo','c_active','c_home')->get();
@@ -23,7 +30,7 @@ class AdminCategoryController extends Controller
     public function store(RequestCategory $requestCategory)
     {
         $this->isertOrUpdate($requestCategory);
-        return redirect()->back()->with('success','Thêm danh mục thành công');;
+        return redirect()->back()->with('success','Thêm danh mục thành công');
     }
 
     public function edit($id)
@@ -34,8 +41,8 @@ class AdminCategoryController extends Controller
 
     public function update($id,RequestCategory $requestCategory )
     {
-        $this->isertOrUpdate($requestCategory,$id);   
-        return redirect()->back()->with('info','Thêm danh mục thành công');;
+        $this->isertOrUpdate($requestCategory,$id);
+        return redirect()->back()->with('info','Cập nhật danh mục thành công');
     }
 
     public function isertOrUpdate(RequestCategory $requestCategory, $id = '')
@@ -56,22 +63,33 @@ class AdminCategoryController extends Controller
        if ($action) {
            switch ($action) {
                case 'delete':
-                   $actionCategory->delete();
-                   return redirect()->back()->with('danger','Xóa danh mục thành công');
+                    //kiểm tra xem danh mục có sản phẩm hay không
+                    //lấy id sản phẩm
+                    $id_products = Product::where('pro_category_id',$id)->pluck('id');
+                   // dd($id_products);
+                    if(count($id_products) == 0)
+                    {
+                        $actionCategory->delete();
+                        return redirect()->back()->with('danger','Xóa danh mục thành công');
+                    }
+                    else
+                    {
+                        return redirect()->back()->with('warning','Danh mục có sản phẩm! bạn không thể xóa');
+                    }
                    break;
                case 'active':
                    $actionCategory->c_active = $actionCategory->c_active == 1 ? 0 : 1;
                    //$actionCategory->c_active = 1 or 0
                    $actionCategory->save();
                   return redirect()->back()->with('success','Cập nhật thành công');
-                       break; 
+                       break;
                case 'home':
                  $actionCategory->c_home = $actionCategory->c_home == 1 ? 0 : 1;
                  $actionCategory->save();
                  return redirect()->back()->with('success','Cập nhật thành công');
-                  break;           
+                  break;
            }
        }
-      
+
     }
 }
